+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 1.4.0 - 2023-05-31
+### Added
+- Added new dereferencing function for up to date openapi-spec and jsonschema-spec libraries.
+
+### Changed
+- Updated openapi-spec to the latest version. This changes create_spec to Spec.from_file().
+- spec object created is different, less proprietary objects.
+- $refs are lazy dereferenced, so I wrote a dereferencer to output a ready spec that we can pickle.
+- Updates to accommodate these changes.
+- Found one bug in jsonschema-spec that I created an issue about, fixed it locally by overwriting a class method.
+
+### Removed
+- Spec are no longer created at initialization, only need to be loaded. 7 seconds -> .04 seconds.
+
 
 ## 1.3.4 - 2023-05-16
 ### Added